refactor: move config models into config dir
