+++ jarvis_sdk_node/src/sdk/model/config/__test__/application_agent.test.ts
@@ -1,6 +1,6 @@
-import { StringValue } from '../../../grpc/google/protobuf/wrappers';
-import { SdkError, SdkErrorCode } from '../../error';
-import { Utils } from '../../utils/utils';
+import { StringValue } from '../../../../grpc/google/protobuf/wrappers';
+import { SdkError, SdkErrorCode } from '../../../error';
+import { Utils } from '../../../utils/utils';
 import { ApplicationAgent } from '../application_agent';
 
 describe('deserialize', () => {
