+++ jarvis_sdk_node/src/sdk/model/config/tenant.ts
@@ -1,9 +1,9 @@
-import { SdkErrorCode, SdkError } from '../error';
+import { SdkErrorCode, SdkError } from '../../error';
 import {
   CreateTenantResponse,
   ReadTenantResponse,
-} from '../../grpc/indykite/config/v1beta1/config_management_api';
-import { Utils } from '../utils/utils';
+} from '../../../grpc/indykite/config/v1beta1/config_management_api';
+import { Utils } from '../../utils/utils';
 
 export class Tenant {
   constructor(
